attempt to reduce CPU overload
As it is right now, if the FrameQueue is full, it means we have 200-ish
frames waiting to be used. This means that even if we used one frame per
sample (hopsize = 1), it would take about 4ms (1000*200/48000) to empty
the FrameQueue. And much more for more likely values of hopsize.

So it seems reasonable to me to have a loop every 1ms (i.e. at 1000Hz)
instead of every 10µs (i.e. at 100kHz).

Besides it reduces my CPU usage by 2 :-)